requirements.txt
fastapi
uvicorn[standard]
pymysql
requests
python-dotenv
jinja2
python-multipart
email-validator
watchfiles